LayoutTests/platform/chromium/permissionclient/storage-permission-expected.txt

This test verifies that all access to localStorage can be blocked
Length is 0
Value for FOO is null
Length is 1
Value for FOO is BAR
Key for index 0 is FOO
Disabling localStorage access.
Caught exception trying to get frame localStorage: Error: SECURITY_ERR: DOM Exception 18
